show storage in profile

+++ src/icdrive_assets/src/components/CenterPortion/Home/GridView.jsx
@@ -23,6 +23,7 @@ const GridView = () => {
   const [ShareLoadingFlag, setShareLoadingFlag] = React.useState(false);
   const [removeFlag, setRemoveLoadingFlag] = React.useState(false);
   const [PublicLoadingFlag, setPublicLoadingFlag] = React.useState(false);
+  const [deletingFlag, setDeletingFlag] = React.useState(false);
   const userName = React.useRef('');
 
   const handleDownload = async () => {
@@ -34,6 +35,7 @@ const GridView = () => {
     for (let i = 0; i < temp.length; i += 1) {
       if (temp[i].fileId === fileObj.current.fileId) {
         temp[i].marked = !temp[i].marked;
+        break;
       }
     }
     dispatch(filesUpdate(temp));
@@ -41,8 +43,14 @@ const GridView = () => {
   };
 
   const handleDelete = async () => {
-    await deleteFile(fileObj.current);
-    dispatch(refreshFiles(true));
+    if(!deletingFlag){
+      setDeletingFlag(true);
+      await deleteFile(fileObj.current);
+      dispatch(refreshFiles(true));
+      setDeletingFlag(false);
+    } else{
+      message.info('Please wait for previous file to delete!!!');
+    }
   };
 
   const handleView = async () => {
