+++ src/icdrive_assets/src/components/CenterPortion/Marked/ListViewMarked.jsx
@@ -26,6 +26,7 @@ const ListViewMarked = () => {
   const [ShareLoadingFlag, setShareLoadingFlag] = React.useState(false);
   const [removeFlag, setRemoveLoadingFlag] = React.useState(false);
   const [PublicLoadingFlag, setPublicLoadingFlag] = React.useState(false);
+  const [deletingFlag, setDeletingFlag] = React.useState(false);
   const userName = React.useRef('');
 
   // Functions
@@ -47,7 +48,8 @@ const ListViewMarked = () => {
     const temp = [...files];
     for (let i = 0; i < temp.length; i += 1) {
       if (temp[i].fileId === record.fileId) {
-        temp[i].marked = !temp[i].marked;
+        temp[i].marked = false;
+        break;
       }
     }
     dispatch(filesUpdate(temp));
@@ -55,8 +57,14 @@ const ListViewMarked = () => {
   };
 
   const handleDelete = async (record) => {
-    await deleteFile(record);
-    dispatch(refreshFiles(true));
+    if(!deletingFlag){
+      setDeletingFlag(true);
+      await deleteFile(record);
+      dispatch(refreshFiles(true));
+      setDeletingFlag(false);
+    } else{
+      message.info('Please wait for previous file to delete!!!');
+    }
   };
 
   const handleView = async (record) => {
